Changes to P2
I made changes to P2 according to the new schema that was created
during project 3. However, some on of my headers in the Additional
Information is not being placed correctly and I don’t know why. Also
some of my text in the Job tag is being cut. I hope my marks don’t get
affected because of that